Nearest stack search: display and select from neighbors
This is an initial attempt at allowing users to switch between close
stacks. A window is displayed in the bottom-left of the screen listing
the nearby-stacks to the one selected, and these are highlighted. A
user can mouse-over the stack names in the table and see the FOV
highlighted (filled with partly-opaque white), and switch to the stack.