Fix "nearest selection"
The routine to calculate separation of two coordinates had the comment
"let's ignore rounding errors", which turned out to not be the best
idea, since it failed when the point being searched was identical to
a location. This then causes the "select nearby sources" code to
fail (in that it would select nearby sources rather than the correct one),
and could have been one of the reasons why stack selection occasionally
acted strangely. The coordinate calculation has been improved, but
may still not be ideal.

There has been some code clean up in the source selection and searching
areas.