Lazy creation of sources (CSC 2.0, 1.1, and XMM)
Rather than create the WWT annotation objects for all sources when the
data has been loaded, this now creates them on the fly (when a user says
"show sources in this area").

This reduces the "download" time at the expense of increased run time each
time a "show sources" command is issued. However, it is likely that the
latter is imperceptible (apart from the most extreme cases).

It also has the advantage of making the color/size changing more responsive,
since a much-smaller number of annotations have to be changed.

This should reduce memory usage (not tested).

It also paves the way to reading these catalogs "on demand" (in a similar
manner to ESA sky) rather than "all in one", if we so desire to go in
this direction.