Replace screen shot of stack info with actual pane
The help page now uses an actual "pane" rather than a screen shot
of the stack info pane. This means that changes to the display code
should automatically get reflected on the help panel.

To simplify matters (and because it will have to be done for the source
properties), the example stack uses hard-coded values, rather than use
the current values. This data will need to be kept up to date with
schema and data changes.