Switch to fontawesome icons
Switch from glyphicons, which were freely available at the time I
obtained them but no longer are, to the free version of fontawesome.
This is to ensure a "consistent" styling to the icons.

I have also removed the use of CSS styling to fake the close and
hide/show icons, replacing them with fontawesome icons.