Bring help/about text up to date
This also adds in the idea of the about/credits page titles always being
visible, so you can close the page whenever you want. I think it will
probably be better to make the content a second div that can be scrolled
rather than this approach, but that is to be tested.