Fix toggle settings button and start pos of settings/preselected
The toggle settings button did not properly recognize when the pane
had been closed by the "close-pane" button, which lead to it failing
to re-display the settings window the next time the show button was
selected.

The settings and preselected panes now open up close to the mouse
position (just to the right and above a little bit) to make the UI
a bit-more reactive.