Attempt to center the select buttons, but fail
Switched to flex to layout the buttons in the user panel in the hope
that this could be used to center the select widgets. It's not clear
it's possible without removing the button styling (or, at least, greatly
reworking it), and as the unstyled version looks significantly different,
leave for now (although keep the change to flex).