Fix settings.
I recently added in logic to only call the settings "on change" handler
if the value is not the default. However, it needs to be set to ensure
the correct setting (in particular for those where the default does not
match the default WWT behavior), so I have removed this. I don't think
this should break anything else ...