Improve positioning of the source-properties panel
With recent changes, the source-properties (color and size) panel
was appearing by the coordinate display in the control panel, rather
than by the toggle button. This improves this, by opening the panel
at the y position of the event. Perhaps it should actually query for
the toggle button and align with the top of that.