Do not announce message-sent status when using notify rather than notifyAll
The notify SAMP call doesn't report whether the client has responded, so
we can't easily determine this information. As the user has selected the
particular application this should be okay from a UI perspective.

Note that the response from notifyAll doesn't guarantee that a client
has received the message, just that one was sent, so it's not a great
indicator, and was added before the recent push to make the SAMP
functionality "more obvious".