Position clipboard so it is just offset from the click
This moves the positioning logic out into a separate routine so it can
be used for other elements (e.g. tooltips, but this is not implemented yet).
The idea is that the click +/- buffer (5 pixels) is used to determine
whether the element top/bottom and left/right is placed at the position
(although the actual positioning is always done using the left and top
CSS values).