Initial support for faceting of the displayed sources
Allow the source display to be toggled: that is, subset the selected
set of sources. The UI for this simple faceting is confusing, and the
choice of options are

a) not well labelled (needs to be re-written before release)
b) very limited
c) not used in ADQL queries

The subset is reflected in the 'plot properties' code.

Note that filtering the sources (either by turning off or on a toggle)
invalidates the "nearest source" and "selected source" window, so they
are removed if displayed. This is not always needed, but is the easiest
to implement and probably explain.