TOGGLE: Initial support for source toggles
Allow the source display to be toggled: that is, subset the selected
set of sources. The UI for this simple faceting is confusing, and the
choice of options are

a) not well labelled (needs to be re-written before release)
b) very limited

The plot button doesn't react to the subset.