Add a panel describing source selection
This panel is used to provide a more-obvious place for the plot and
export-to-samp actions, although the latter is not set up yet.

The plot-source-properties and export-to-samp actions have been moved
from the control panel to this source panel.

The plot code now creates the data for plotting "on the fly" rather
than when the sources are initially selected.