Start work on supporting export to specific SAMP clients
This adds a widget that lets the user control which SAMP client the
search is sent to, *including* a copy-to-clipboard option which does
not require SAMP to be running.

This is only implemented for the "source properties near" search, and
is only partially working. The choice of targets does not react to
changes (e.g. clients being added or removed, or connecting to a hub),
and the visual appearance needs work.