Ensure a sensible selection for SAMP client list when it changes
The previous behavior was to always to pick the first element (copy to
clipboard) whenever the SAMP client list changed. The idea now is that
the previous selection is used if it still present, otherwise use (in
decreasing order of priority)

  - client name (single client only) or 'all clients'
  - copy to clipboard

Export symbols and a function from wwtsamp to define the value field of
the options, rather than hard-code the string values.