Be explicit about what data is exported
This uses the same look for the "selected stack" and "selected source"
panels as for the "selected set of sources" for the export section, in
that it is now explicit about what is being exported. Unlike the
"set of sources" case, there is no choice for the stack and single-source
cases.

The code has been refactored to reduce duplication.