Improve UX when user has selected 'find target' for SAMP export
During user testing, it was found that after a 'find target' selection
the export button could still result in a copy-to-clipboard action.
This is because the registration and UI update for the SAMP hub is
asynchronous, and can take a significant amount of time. For the moment
the approach is to ignore the export button while this is happening,
but a better approach may be:

a) disable the button for the duration
b) switch to 'all clients' for the duration