Try to stop "bouncing" of tooltip (appear, disappear, then reappear)
Ensure that a tooltip can not appear too soon after it has been displayed,
to try and stop "bouncing" as seen in the current code base. I am not
sure what causes the bounce (I wonder if the mouse appears to re-enter
the element when the tooltip disappears which re-triggers things, but
the offset should stop this happening more than once), so try and limit
a particular tooltip so that it can't happen too often.