When turning off SAMP support, turn off SAMP targets in client lists
The logic of what "turning off" SAMP support is not 100% clear in the
code, since we do not destroy the SAMP connection, just stop the
polling. This commit adds an explicit function (isRunning) to
indicate whether we are "allowed" to make SAMP calls or not.

Note that even when wwtsamp.isRunning() is false it is still possible
to make SAMP calls, but the assumption is that you are not going to.
The code could be enhanced to enforce this restriction but let's wait
to see if it is needed.