Use https URLs for tile access
The WWT control comes with support for most of the wavebands we
use so the WTML file could be greatly simplified, *EXCEPT* that
at the moment the inbuilt records use http rather than https
URLs, so I am hard-coding versions for us. To avoid tripping
over the hard-coded values (since the name field does not guarantee
it will be over-written) I have tweaked the names to start with
'X ' as a work-around.

That is

  wwtlib.WWTControl.imageSets.map(d => d.url.startsWith('https'))

returns a bunch of False values.