Improve point selection and WTML use
We store the fill color of circles (as having to use a combined
opacity + color setting to work around issues) so that we can
re-create the setting when selecting sources (in particular
when restoring a selected source to a non-selected source).

Move to the inbuilt image sets to see if this works better
(i.e. get https access)