Read in names/obis via AJAX
This makes it easier to write code that is somewhat version-agnostic.

We do now create different stack output for CSC 2.1 compared to
2.0, as we can say "this has changed and these are new", "this is
a new stack", or "this is unchanged from CSC 2.0".

There's some changes to make the code more generic, but work is
still needed.

The WWT2.1 interface is provided as a starter. Much more work is
needed to be useful.