Allow stack selection via URL parameter
Adding stackid=acisfJxxx_001 (for a valid stack) will select the
stack and bring up the stack information pane. It does not move to
the stack, so you would still need to use the ra, dec, and zoom
parameters for that.

The bookmark facility will now include the stackid parameter if
a stack is selected.