Use a spinner to mark the whole initialization process
I had been using a spinner just to mark data being downloaded,
but there's a relatively long time when we appear to be waiting,
for WWT to call the "I am ready" function. So we now start the
spinner when initialization starts, and stop once the wwtReadyFunc
is finished.

I will note that this may not work well if the re-initialize logic
is triggered, but I'm hoping this isn't needed anymore (that is,
the underlying WWT issue has been resolved).

Other changes are that some data is now downladed earlier, in
a hope to reduce the start up time slightly.