Get source properties for CSC 2.1
This meant re-working the source properties for CSC 2.0,
as well as finding a number of places where we had hard-coded
2.0 and needed to either replace it with 2.1 or make it depend
on a JS value.

The catalog properties work was not too bad, but there is a
number of places where we could perhaps just have a "csc2" version
rather than "csc20" or "csc21".

In particular, the data loading currently is almost identical and
could be handled more generically.