wwt: save XMM settings, save CSC color, tweak opacity
CSC sources now use the same fill and line colors.
We save and restore CSC (>= 2) and XMM size, opacity, and colors.
We do not do this for CSC 1.1 but could.

It looks like we share the same "namespace" for settings with
wwt and wwt21 pages, which is not what I expected.